+++ packages/atlas/src/views/viewer/ChannelView/ChannelView.styles.ts
@@ -1,12 +1,9 @@
-import { css } from '@emotion/react'
 import styled from '@emotion/styled'
 
 import { Tabs } from '@/components/Tabs'
 import { Text } from '@/components/Text'
 import { Button } from '@/components/_buttons/Button'
-import { IconButton } from '@/components/_buttons/IconButton'
 import { ChannelLink } from '@/components/_channel/ChannelLink'
-import { TextField } from '@/components/_inputs/TextField'
 import { SkeletonLoader } from '@/components/_loaders/SkeletonLoader'
 import { cVar, media, oldColors, sizes, transitions } from '@/styles'
 
@@ -14,18 +11,6 @@ const SM_TITLE_HEIGHT = '44px'
 const TITLE_HEIGHT = '51px'
 const SM_SUBTITLE_HEIGHT = '24px'
 const SUBTITLE_HEIGHT = '27px'
-
-const activeUnderline = css`
-  &::after {
-    content: '';
-    display: block;
-    position: absolute;
-    width: 100%;
-    height: 4px;
-    background-color: ${oldColors.blue[500]};
-    bottom: -${sizes(3)};
-  }
-`
 
 export const TitleSection = styled.div`
   margin-top: -${sizes(6)};
@@ -164,22 +149,6 @@ export const TabsContainer = styled.div`
   }
 `
 
-type SearchContainerProps = {
-  isOpen?: boolean
-}
-export const SearchContainer = styled.div<SearchContainerProps>`
-  display: flex;
-  grid-area: search;
-  align-items: center;
-  margin: ${sizes(8)} 0 ${sizes(2)} 0;
-  position: relative;
-
-  ${media.sm} {
-    grid-area: initial;
-    margin: 0;
-  }
-`
-
 export const StyledTabs = styled(Tabs)`
   grid-area: tabs;
   border-bottom: solid 1px ${oldColors.gray[700]};
@@ -190,49 +159,6 @@ export const StyledTabs = styled(Tabs)`
   }
 `
 
-type TextFieldProps = {
-  isOpen?: boolean
-  isSearching?: boolean
-}
-export const StyledTextField = styled(TextField)<TextFieldProps>`
-  transition: all ${transitions.timings.regular} ${transitions.easing};
-  will-change: max-width;
-  align-items: center;
-  position: relative;
-
-  ${media.sm} {
-    max-width: ${({ isOpen }) => (isOpen ? '200px' : '0px')};
-  }
-
-  ${({ isSearching }) => isSearching && activeUnderline}
-
-  input {
-    height: 40px;
-    padding: 10px 16px 10px 42px;
-    caret-color: ${oldColors.blue[500]};
-    font: ${cVar('typographyDesktopT200')};
-    letter-spacing: ${cVar('typographyDesktopT200LetterSpacing')};
-    text-transform: ${cVar('typographyDesktopT200TextTransform')};
-
-    ${media.sm} {
-      ${({ isOpen }) => isOpen === false && 'border: none !important'};
-    }
-
-    :focus {
-      border: 1px solid ${oldColors.white};
-    }
-
-    ::-webkit-search-cancel-button {
-      /* stylelint-disable-next-line property-no-vendor-prefix */
-      -webkit-appearance: none;
-    }
-  }
-`
-
-type SearchButttonProps = {
-  isSearching?: boolean
-  isOpen?: boolean
-}
 export const NotFoundChannelContainer = styled.div`
   display: flex;
   align-items: center;
@@ -240,13 +166,6 @@ export const NotFoundChannelContainer = styled.div`
   height: calc(100vh - var(--size-topbar-height));
 `
 
-export const SearchButton = styled(IconButton)<SearchButttonProps>`
-  position: absolute;
-
-  ${media.sm} {
-    ${({ isSearching, isOpen }) => isSearching && !isOpen && activeUnderline}
-  }
-`
 export const CollectorsBoxContainer = styled.div`
   display: flex;
   height: 64px;
